Procfs storage (#26)
Adds procfs metric collection, in-memory storage, and aggregation/diff utilities so the master can periodically gather per-process procfs stats from segment agents and retain/compare snapshots over time.

Changes:

- Introduce ProcfsStorage (time-series snapshot storage) plus procfs diff/grouping helpers and tests.
- Add master-side procfs gatherer (ProcfsGatherStorage) and wire it into the background refresh loop + config/app initialization.
- Add aggregation step after refreshing list of sessions. Add to the sessions stats procfs stat with current top5 resource usage (to the `LongRunningGPMetrics` field)

+++ internal/grpc/get_query_info_linux_test.go
@@ -33,12 +33,12 @@ func TestGpPidProcInfo_CurrentProcess(t *testing.T) {
 	assert.NotEmpty(t, info.Cmdline)
 
 	require.NotNil(t, info.ProcStat)
-	assert.Equal(t, int32(pid), info.ProcStat.Pid)
+	assert.Equal(t, int64(pid), info.ProcStat.Pid)
 	assert.NotEmpty(t, info.ProcStat.Comm)
 	assert.NotEmpty(t, info.ProcStat.State)
 
 	require.NotNil(t, info.ProcStatus)
-	assert.Equal(t, int32(pid), info.ProcStatus.Pid)
+	assert.Equal(t, int64(pid), info.ProcStatus.Pid)
 	assert.NotEmpty(t, info.ProcStatus.Name)
 }
